+++ src/ProcessManager.php
@@ -12,7 +12,7 @@ use Symfony\Component\Process\Process;
  * @author BluePsyduck <bluepsyduck@gmx.com>
  * @license http://opensource.org/licenses/GPL-3.0 GPL v3
  */
-class ProcessManager
+class ProcessManager implements ProcessManagerInterface
 {
     /**
      * The number of processes to run in parallel.
@@ -25,6 +25,12 @@ class ProcessManager
      * @var int
      */
     protected $pollInterval;
+
+    /**
+     * The time to delay the start of processes to space them out, in milliseconds.
+     * @var int
+     */
+    protected $processStartDelay;
 
     /**
      * The processes currently waiting to be executed.
@@ -54,11 +60,50 @@ class ProcessManager
      * ProcessManager constructor.
      * @param int $numberOfParallelProcesses The number of processes to run in parallel.
      * @param int $pollInterval The interval to wait between the polls of the processes, in milliseconds.
+     * @param int $processStartDelay The time to delay the start of processes to space them out, in milliseconds.
      */
-    public function __construct(int $numberOfParallelProcesses = 1, int $pollInterval = 100)
-    {
+    public function __construct(
+        int $numberOfParallelProcesses = 1,
+        int $pollInterval = 100,
+        int $processStartDelay = 0
+    ) {
         $this->numberOfParallelProcesses = $numberOfParallelProcesses;
         $this->pollInterval = $pollInterval;
+        $this->processStartDelay = $processStartDelay;
+    }
+
+    /**
+     * Sets the number of processes to run in parallel.
+     * @param int $numberOfParallelProcesses
+     * @return $this
+     */
+    public function setNumberOfParallelProcesses(int $numberOfParallelProcesses)
+    {
+        $this->numberOfParallelProcesses = $numberOfParallelProcesses;
+        $this->executeNextPendingProcess(); // Start new processes in case we increased the limit.
+        return $this;
+    }
+
+    /**
+     * Sets the interval to wait between the polls of the processes, in milliseconds.
+     * @param int $pollInterval
+     * @return $this
+     */
+    public function setPollInterval(int $pollInterval)
+    {
+        $this->pollInterval = $pollInterval;
+        return $this;
+    }
+
+    /**
+     * Sets the time to delay the start of processes to space them out, in milliseconds.
+     * @param int $processStartDelay
+     * @return $this
+     */
+    public function setProcessStartDelay(int $processStartDelay)
+    {
+        $this->processStartDelay = $processStartDelay;
+        return $this;
     }
 
     /**
@@ -116,11 +161,20 @@ class ProcessManager
     protected function executeNextPendingProcess(): void
     {
         if ($this->canExecuteNextPendingRequest()) {
-            list($process, $callback, $env) = array_shift($this->pendingProcessData);
+            $this->sleep($this->processStartDelay);
+
+            [$process, $callback, $env] = array_shift($this->pendingProcessData);
             /* @var Process $process */
             $this->invokeCallback($this->processStartCallback, $process);
             $process->start($callback, $env);
-            $this->runningProcesses[$process->getPid()] = $process;
+
+            $pid = $process->getPid();
+            if ($pid === null) {
+                // The process finished before we were able to check its process id.
+                $this->checkRunningProcess($pid, $process);
+            } else {
+                $this->runningProcesses[$pid] = $process;
+            }
         }
     }
 
@@ -146,16 +200,18 @@ class ProcessManager
 
     /**
      * Checks the process whether it has finished.
-     * @param int $pid
+     * @param int|null $pid
      * @param Process $process
      */
-    protected function checkRunningProcess(int $pid, Process $process): void
+    protected function checkRunningProcess(?int $pid, Process $process): void
     {
         $process->checkTimeout();
         if (!$process->isRunning()) {
             $this->invokeCallback($this->processFinishCallback, $process);
 
-            unset($this->runningProcesses[$pid]);
+            if ($pid !== null) {
+                unset($this->runningProcesses[$pid]);
+            }
             $this->executeNextPendingProcess();
         }
     }
@@ -167,18 +223,19 @@ class ProcessManager
     public function waitForAllProcesses()
     {
         while ($this->hasUnfinishedProcesses()) {
-            $this->sleep();
+            $this->sleep($this->pollInterval);
             $this->checkRunningProcesses();
         }
         return $this;
     }
 
     /**
-     * Sleeps for the next poll.
+     * Sleeps for the specified number of milliseconds.
+     * @param int $milliseconds
      */
-    protected function sleep(): void
+    protected function sleep(int $milliseconds): void
     {
-        usleep($this->pollInterval * 1000);
+        usleep($milliseconds * 1000);
     }
 
     /**